feat(HW-819): Indexer types refactoring
